backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
groq==0.11.0
pydantic>=2.7.4,<3.0.0
pydantic-settings==2.2.1
numpy==1.26.4
scikit-learn==1.4.2
httpx==0.27.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
motor>=3.3.2
